Create pacman cache dir in build.ps1
Otherwise, the build will fail if the directory is not there, see e.g. https://github.com/open62541/open62541/issues/2068.

The problem hasn't shown up on master because the directory is cached, but if that fails for whatever reason (cache deleted...), the build would fail. I noticed this on a fork where I don't cache this directory.