Support functions in custom clipboard providers
Custom clipboard providers can be set using g:clipboard, where
g:clipboard['copy']['+'] was originally meant to be a string with
a command that sets the clipboard.

This commit enables the use of vimscript functions instead. The function
signatures is the same as the clipboard provider itself. A clipboard
provider is expected to store and return a list of lines (i.e. the text)
and a register type (as seen in setreg()).

The cache_enabled attributed is completely ignored if the 'copy'
attribute is set to be a function.