multi signcols: fix bug related to wrapped lines
In wrapped lines, the wrapped parts of a line did not include the extra
columns if they existed. The result was a misdrawing of the wrapped
parts. This fixes the issue, by 1) initializing the signcol counter
back to 0 when we are on a wrap boundary, and 2) allowing for the draw
of spaces in that case.