pvscheck.sh: Ignore V1028 "cast operands, not the result"
This PVS warning, while useful, conflicts with changes done to satisfy
`-Wconversion`.

For example:

    assert(orig_char_len + size - ind_done + line_len >= 0);
    newline = xmalloc((size_t)(orig_char_len + size - ind_done + line_len));

We verified that the result is valid for xmalloc, but PVS warns because
the individual terms may result in nonsense (e.g. ind_done + line_len
could overflow `int`, which is UB).

The PVS warning is valid, but satisfying it would be onerous, and
unfeasible.