:terminal : avoid deleted_lines()
Problem:
--------
Running a command (e.g. `tree /`) that prints a lot of output in
a hidden :terminal buffer, causes spurious redraws in the current
buffer.

Test case:

    nvim -u NORC foo
    :set hidden
    :terminal tree /
    :bp
    <input several lines of text while `tree /` runs in the background>

Solution:
---------
deleted_lines() (calls changed_lines()) is expensive. The cases removed
in this commit appear to be unnecessary; maybe the changed_lines() call
in refresh_screen() is sufficient?

closes #8723
Helped-by: Sandeep.C.R <sandeepcr2@gmail.com>