Add more granular margin setting
Following a rebase and fixing some compilation issues, this change
introduces a new way to specify each of the margins. The format is as
follows: margin=<num>,start:<num>,end:<num>,top:<num>,bottom:<num>. All
values are optional, where the first number passed will set all margins
after which point specific margins can be set explicitly. The margin
value still stands as the number of rows and/or columns in the window
instead of a percentage.