keymap.c/keymap.h: change name K_KCENTER -> K_KORIGIN
This name seems more appropriate as this is similar to what is used by
xev. Alternative name could be K_KBEGIN as this seems to be referenced
by this name here:
https://invisible-island.net/xterm/ctlseqs/ctlseqs.html#h2-PC-Style-Function-Keys

But I fine Begin and Home too similar, and it might induced some
confusion.
The naming looked related to some old keyboard configuration.