Allow running a command after remote attach
gdbserver has a bug which means that it can't actually attach to running
processes on some versions due to sending signals to the wrong PID (not
the process group leader, or something). The workaround is to manually
send kill -TRAP to the PID in order for the initial break to be
regiseterd.

Doing this manually is painful, so we allow the remote object to run a
command after having successfully initialized the engine, which is about
the time you need to kick the trap/breakpoint.

https://sourceware.org/bugzilla/show_bug.cgi?id=18945