runtime: Add bufnr to client request argument.
If we receive the language server response when our currnent buffer is different from we posted request to the languager server, we can't know that we should apply to which buffer.
So I add bufnr to request method argument.
If there is no bufnr argument parameter, the client use current bufnr as argument.