[RFC/WIP] nv_next: do not re-do search unnecessarily
Uses a new "wrapped" flag to skip the additional search when the
position was not changed (but 'wrapscan' wrapped around).

This makes it necessary to change a lot of places to pass this through.
Given that it is similar to tm/timed_out being passed around already, it
might make sense to use a single struct/mask here to be passed around
instead, allowing for easier changes in the future.

Ref/via: https://github.com/neovim/neovim/issues/11243#issuecomment-542949801