shared.lua: make lua nvim stdlib to raise error with traceback
If we get error in shared.lua function, we get message with shared.vim file line number which shows error location.
But what we really want to know is not the location of the error in shared.lua, but the information of where the function in shared.lua was called.
So by giving traceback to errors in shared.lua, more meaningful errors can be displayed.