logging: nvim_log() all levels if 'verbose' > 0
If verbose >= 1 then nvim_log() logs all levels. If verbose == 0 then
nvim_log() only logs ERROR level.

This commit intentionally keeps the old verbose=1 behavior for `:set`,
because that is unlikely to get in the way.  Other verbose=1 behaviors
were moved to verbose=2, so that the editor can be used without a bunch
of noisy messages barfing on the screen.