Add a create_logger(filename) function to lsp/log.lua
The motivation is to be able to use the log functions in plugins without
having it log into the `vim-lsp.log` file.

To keep the interface of `log` as it was before. Importing the `log`
module will still result in a logger that is targeting `vim-lsp.log`,
but it additionally contains a `create_logger(filename)` function that
can be used to create different loggers.

E.g.:

    local log = require('lsp.log')
    log.warn('This logs to vim-lsp.log')

    local mylogger = log.create_logger('myplugin.log')
    mylogger.warn('This logs to myplugin.log')

See https://github.com/mfussenegger/nvim-dap/commit/328fdd9a129f3fdb4d662d03dbdfaf3859b809c7 where this could be used